app/main.py
from fastapi import FastAPI, HTTPException, UploadFile, File, BackgroundTasks
from fastapi.responses import StreamingResponse, JSONResponse
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field
from typing import Optional, List, Dict, Any
from PIL import Image
import io
import numpy as np
import logging
from datetime import datetime
import base64
from concurrent.futures import ThreadPoolExecutor

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Initialize FastAPI app
app = FastAPI(
    title="GAN Model API",
    description="Production API for Generative Adversarial Network inference",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Thread pool for CPU-bound operations
executor = ThreadPoolExecutor(max_workers=4)

# Global state for model (replace with your model loading logic)
model_state = {
    "generator_loaded": False,
    "discriminator_loaded": False,
    "device": "cpu",
    "model_version": "1.0.0"
}

# ==================== Pydantic Models ====================

class GenerateRequest(BaseModel):
    num_images: int = Field(default=1, ge=1, le=10, description="Number of images to generate")
    latent_vector: Optional[List[float]] = Field(default=None, description="Custom latent vector")
    seed: Optional[int] = Field(default=None, description="Random seed for reproducibility")
    return_format: str = Field(default="base64", description="Return format: base64 or url")
    image_size: Optional[int] = Field(default=256, description="Output image size")

class GenerateResponse(BaseModel):
    images: List[str]
    latent_vectors: Optional[List[List[float]]] = None
    generation_time: float
    num_images: int
    timestamp: str

class InterpolateRequest(BaseModel):
    latent_vector_1: List[float]
    latent_vector_2: List[float]
    steps: int = Field(default=5, ge=2, le=20, description="Number of interpolation steps")
    return_format: str = Field(default="base64")

class DiscriminateRequest(BaseModel):
    image_base64: str

class DiscriminateResponse(BaseModel):
    score: float
    is_real_probability: float
    confidence: float
    timestamp: str

class HealthResponse(BaseModel):
    status: str
    device: str
    generator_loaded: bool
    discriminator_loaded: bool
    model_version: str
    uptime_seconds: float
    timestamp: str

class BatchGenerateRequest(BaseModel):
    batch_size: int = Field(default=10, ge=1, le=100)
    latent_vectors: Optional[List[List[float]]] = None
    callback_url: Optional[str] = Field(default=None, description="Webhook URL for completion")

class ModelConfigResponse(BaseModel):
    latent_dim: int
    image_size: int
    image_channels: int
    supported_formats: List[str]

# ==================== Helper Functions ====================

def encode_image_to_base64(image: Image.Image) -> str:
    """Convert PIL Image to base64 string"""
    buffered = io.BytesIO()
    image.save(buffered, format="PNG")
    img_str = base64.b64encode(buffered.getvalue()).decode()
    return img_str

def decode_base64_to_image(base64_str: str) -> Image.Image:
    """Convert base64 string to PIL Image"""
    img_data = base64.b64decode(base64_str)
    image = Image.open(io.BytesIO(img_data))
    return image

# ==================== Core API Endpoints ====================

@app.get("/", tags=["Root"])
async def root():
    """Root endpoint with API information"""
    return {
        "message": "GAN Model API",
        "version": "1.0.0",
        "docs": "/docs",
        "health": "/health"
    }

@app.get("/health", response_model=HealthResponse, tags=["Health"])
async def health_check():
    """Health check endpoint for monitoring"""
    return HealthResponse(
        status="healthy",
        device=model_state["device"],
        generator_loaded=model_state["generator_loaded"],
        discriminator_loaded=model_state["discriminator_loaded"],
        model_version=model_state["model_version"],
        uptime_seconds=0.0,  # Implement actual uptime tracking
        timestamp=datetime.utcnow().isoformat()
    )

@app.get("/config", response_model=ModelConfigResponse, tags=["Configuration"])
async def get_model_config():
    """Get model configuration and capabilities"""
    return ModelConfigResponse(
        latent_dim=512,  # Replace with your model's latent dimension
        image_size=256,  # Replace with your model's output size
        image_channels=3,
        supported_formats=["base64", "url", "bytes"]
    )

# ==================== Generation Endpoints ====================

@app.post("/generate", response_model=GenerateResponse, tags=["Generation"])
async def generate_images(request: GenerateRequest):
    """
    Generate images using the GAN generator
    
    - **num_images**: Number of images to generate (1-10)
    - **latent_vector**: Optional custom latent vector
    - **seed**: Optional random seed for reproducibility
    - **return_format**: Output format (base64, url)
    """
    start_time = datetime.utcnow()
    
    try:
        # TODO: Implement your generation logic here
        # Example: generated_images = your_generator.generate(request.num_images, request.latent_vector)
        
        # Placeholder response
        images = []
        latent_vectors = []
        
        for i in range(request.num_images):
            # Replace with actual generation
            dummy_image = Image.new('RGB', (256, 256), color=(73, 109, 137))
            images.append(encode_image_to_base64(dummy_image))
            latent_vectors.append([0.0] * 512)  # Placeholder latent vector
        
        generation_time = (datetime.utcnow() - start_time).total_seconds()
        
        return GenerateResponse(
            images=images,
            latent_vectors=latent_vectors,
            generation_time=generation_time,
            num_images=request.num_images,
            timestamp=datetime.utcnow().isoformat()
        )
        
    except Exception as e:
        logger.error(f"Generation error: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Generation failed: {str(e)}")

@app.post("/generate/batch", tags=["Generation"])
async def batch_generate(request: BatchGenerateRequest, background_tasks: BackgroundTasks):
    """
    Generate large batches of images asynchronously
    Returns a job ID for tracking progress
    """
    job_id = f"job_{datetime.utcnow().timestamp()}"
    
    # TODO: Implement batch generation with background tasks
    # background_tasks.add_task(process_batch_generation, job_id, request)
    
    return {
        "job_id": job_id,
        "status": "queued",
        "batch_size": request.batch_size,
        "message": "Batch generation started",
        "status_url": f"/status/{job_id}"
    }

# ==================== Model Management Endpoints ====================

@app.post("/model/load", tags=["Model Management"])
async def load_model(model_path: str, model_type: str = "generator"):
    """
    Load or reload model from checkpoint
    Requires appropriate permissions in production
    """
    try:
        # TODO: Implement model loading
        # load_checkpoint(model_path, model_type)
        
        if model_type == "generator":
            model_state["generator_loaded"] = True
        elif model_type == "discriminator":
            model_state["discriminator_loaded"] = True
        
        return {
            "status": "success",
            "model_type": model_type,
            "model_path": model_path,
            "timestamp": datetime.utcnow().isoformat()
        }
        
    except Exception as e:
        logger.error(f"Model loading error: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Model loading failed: {str(e)}")

@app.get("/metrics", tags=["Monitoring"])
async def get_metrics():
    """
    Get API metrics for monitoring
    """
    return {
        "total_requests": 0,  # Implement request counter
        "avg_generation_time": 0.0,
        "error_rate": 0.0,
        "cache_hit_rate": 0.0,
        "timestamp": datetime.utcnow().isoformat()
    }

# ==================== Startup/Shutdown Events ====================

@app.on_event("startup")
async def startup_event():
    """Initialize models and resources on startup"""
    logger.info("Starting GAN API server...")
    # TODO: Load your models here
    # model_manager.load_generator("path/to/generator.pth")
    # model_manager.load_discriminator("path/to/discriminator.pth")
    logger.info("GAN API server ready")

@app.on_event("shutdown")
async def shutdown_event():
    """Cleanup resources on shutdown"""
    logger.info("Shutting down GAN API server...")
    executor.shutdown(wait=True)
    logger.info("GAN API server stopped")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000, log_level="info")
